CPU Multithreading Simulator
This was my final project in Operating Systems. Many, many sleepless nights went into getting this program to the point it's at. Unfortunately, the final calculations are off; however, I believe the faults present are with my math skills and not so much with the overall functions. The program correctly handles and passes the bursts, which leads me to question my methods of calculating the final data.